+++ paddorch/tensor.py
@@ -40,7 +40,7 @@ class Tensor(paddle.Tensor  ):
         elif isinstance(args[0],Iterable):
             args=list(args)
             if isinstance(args[0][0],int):
-                args[0] = np.array(args[0]).astype("int64")
+                args[0] = np.array(args[0]).astype("int32")
             else:
                 args[0]=np.array(args[0]).astype("float32")
             super(Tensor, self).__init__(*args, **kwargs)
@@ -257,14 +257,14 @@ class Tensor(paddle.Tensor  ):
         if isinstance(args[0],paddle.Tensor):
             dtype=str(args[0].dtype)
             if "64" in dtype:
-                dtype="int64"
+                dtype="int32"
             elif "32" in dtype:
                 dtype = "float32"
             else:
                 return self
         else:
             dtype=str(args[0])
-        if dtype=="int64":
+        if dtype=="int32":
             return self.long()
         elif dtype=="float32":
             return self.float()
@@ -341,7 +341,7 @@ class Tensor(paddle.Tensor  ):
                 elif key.dtype==paddle.fluid.core.VarDesc.VarType.INT32 or key.dtype==paddle.fluid.core.VarDesc.VarType.INT64:
                     return key
                 else:
-                    return key.astype("int64")
+                    return key.astype("int32")
             if isinstance(key,int):
                 return paddorch.LongTensor(np.array([key]))
             if isinstance(key,list):
@@ -485,7 +485,7 @@ class Tensor(paddle.Tensor  ):
 
 
     def index_select(self,dim, index):
-        index=index.astype("int64")
+        index=index.astype("int32")
 
         ret= convertTensor(paddle.index_select(self,index=index, axis=dim))
         return  ret
