+++ paddorch/test/scatter_test.py
@@ -1,7 +1,7 @@
 import paddle
 
 a=paddle.zeros([1]).astype("float32")
-b=paddle.zeros([5]).astype("int64")
+b=paddle.zeros([5]).astype("int32")
 c=paddle.arange(5,10).astype("float32")
 
 d=paddle.scatter(a,b,c,overwrite=False)
