+++ paddorch/utils/dlpack.py
@@ -21,16 +21,16 @@ def from_dlpack(dlpack):
     tensor_from_dlpack = fluid.core.from_dlpack(dlpack)
     place=tensor_from_dlpack._place()
     if  True:# "win" in platform: # CPU env
-        if "int64" in str(tensor_from_dlpack):
-            return paddorch.convertTensor(paddle.to_tensor(np.array(tensor_from_dlpack),dtype="int64"))
+        if "int32" in str(tensor_from_dlpack):
+            return paddorch.convertTensor(paddle.to_tensor(np.array(tensor_from_dlpack),dtype="int32"))
         else:
             return paddorch.Tensor(paddle.to_tensor(np.array(tensor_from_dlpack)))
     else:
         with paddle.fluid.dygraph.guard(place=place):
             tensor_from_dlpack.__class__=paddle.fluid.LoDTensor
             ret= paddle.Tensor( tensor_from_dlpack)
-            if "int64" in str(tensor_from_dlpack):
-                ret=paddle.to_tensor(ret,dtype="int64")
+            if "int32" in str(tensor_from_dlpack):
+                ret=paddle.to_tensor(ret,dtype="int32")
             tensor_from_dlpack.__class__=paddle.fluid.core_avx.Tensor
         return ret
